Refactored discovery.go into separate files - Solves issues #37 - reorganized methods into new files - renamed some comments/methods
